moved newEvent compent to child of EventList component along with state and functions
