+++ client/src/components/EventList/index.js
@@ -3,9 +3,42 @@ import "materialize-css/dist/css/materialize.min.css";
 import "./style.css";
 import { Table } from "react-materialize";
 import EventListItem from "./EventListItem"
+import NewEvent from "./NewEvent";
 import Axios from "axios";
 
 export class EventList extends React.Component {
+
+    state = {
+        date: "",
+        title: "",
+        contact: [],
+      };
+    
+      //when we click enter new event, use state as new event and send to db
+      enterNewEvent = x => {
+        x.preventDefault();
+    
+        let newEvent = {
+          date: this.state.date,
+          title: this.state.title,
+          contact: this.state.contact,
+        };
+    
+        Axios.post("api/user/" + this.props.user._id + "/events", newEvent).catch(
+          error => {
+            console.log(error.response);
+          }
+        );
+      };
+    
+      //set state as user enters event info
+      newEvent = event => {
+        const { name, value } = event.target;
+    
+        this.setState({
+          [name]: value,
+        });
+      };
 
     //When user clicks delete an event
   handleDeleteEventClick = id => {
@@ -53,6 +86,11 @@ export class EventList extends React.Component {
                 {this.renderEventList()}
                 </tbody>
             </Table>
+            <br/>
+            <NewEvent
+                  handleNewEvent={this.newEvent}
+                  enterNewEvent={this.enterNewEvent}
+                />
         </div>
         )
     }
