+++ client/src/pages/eventPage.js
@@ -3,41 +3,8 @@ import { Col, Row, Container } from "../components/Grid";
 import Sidebar from "../components/Sidebar";
 import { EventList } from "../components/EventList";
 import { Redirect } from "react-router-dom";
-import NewEvent from "../components/NewEvent";
-import Axios from "axios";
 
 class EventPage extends Component {
-  state = {
-    date: "",
-    title: "",
-    contact: [],
-  };
-
-  //when we click enter new event, use state as new event and send to db
-  enterNewEvent = x => {
-    x.preventDefault();
-
-    let newEvent = {
-      date: this.state.date,
-      title: this.state.title,
-      contact: this.state.contact,
-    };
-
-    Axios.post("api/user/" + this.props.user._id + "/events", newEvent).catch(
-      error => {
-        console.log(error.response);
-      }
-    );
-  };
-
-  //set state as user enters event info
-  newEvent = event => {
-    const { name, value } = event.target;
-
-    this.setState({
-      [name]: value,
-    });
-  };
 
   render() {
     if (this.props.user.firstName === "George") {
@@ -56,11 +23,6 @@ class EventPage extends Component {
               <Col>
                 <EventList
                   user={this.props.user}
-                />
-                <br />
-                <NewEvent
-                  handleNewEvent={this.newEvent}
-                  enterNewEvent={this.enterNewEvent}
                 />
               </Col>
             </Row>
